requirements.txt
crewai>=0.140.0
crewai-tools>=0.49.0
pypdf>=5.0.0
chromadb>=0.5.23
python-dotenv>=1.0.0
pydantic>=2.8.0
markdown>=3.6
beautifulsoup4>=4.12.0
openai>=1.40.0
tiktoken>=0.8.0
numpy>=2.0.0
pandas>=2.2.0